Reduce classes
